Moved implementation of template class Data to header, not being compiled on itself, Implemented method countAfterpulses() in class DataProcessor, Reworked inclusion of Data template in Event, DriftTimeSpectrum and RtRelation, implemented RtRelation constructors, reworked script to compile unit tests

+++ src/DriftTimeSpectrum.cpp
@@ -9,7 +9,7 @@
 
 using namespace std;
 
-DriftTimeSpectrum::DriftTimeSpectrum(unique_ptr<array<uint16_t,800>> spect,unsigned int nEntries) : Data(move(spect))
+DriftTimeSpectrum::DriftTimeSpectrum(unique_ptr<array<uint32_t,800>> spect, unsigned int nEntries) : Data(move(spect))
 {
 	m_entries = nEntries;
 }
